add authorization token and update users routes

+++ src/app/controllers/users.js
@@ -1,15 +1,8 @@
 // import { params, validationResult } from 'express-validator';
-import { param, validationResult } from 'express-validator';
+import { validationResult } from 'express-validator';
+import bcrypt from 'bcrypt';
 import UserService from '../services/userService';
-
-const validation = method => {
-  switch (method) {
-    case 'getUserById':
-      return [param('userId', 'UserId should be an integer.').isInt()];
-    default:
-      console.log('Default Validation!');
-  }
-};
+import { generateToken } from '../../config/auth';
 
 const getUserById = async (req, res) => {
   try {
@@ -32,6 +25,11 @@ const getUserById = async (req, res) => {
 
 const getUserByCpf = async (req, res) => {
   try {
+    const inputErrors = validationResult(req);
+    if (!inputErrors.isEmpty()) {
+      return res.status(422).json({ inputErrors: inputErrors.array() });
+    }
+
     const response = await UserService.getUserByCpf(req.body.cpf);
     if (response) {
       return res.status(200).send(response);
@@ -40,21 +38,45 @@ const getUserByCpf = async (req, res) => {
     console.log('User with this cpf has not been found');
     return res.status(404).send('User not found!');
   } catch (err) {
-    return res.status(500).send('Something wrong when tried to find a user searching by cpf!');
+    return res.status(500).send('Something wrong when tried to find an user searching by cpf!');
+  }
+};
+
+const getAllUsers = async (req, res) => {
+  try {
+    const response = await UserService.getAllUsers();
+    if (response) {
+      return res.status(200).send(response);
+    }
+
+    console.log('Users not found!');
+    return res.status(404).send('Users not found!');
+  } catch (e) {
+    return res.status(500).send('Something wrong when tried to find all users!');
   }
 };
 
 const createUser = async (req, res) => {
   try {
+    const inputErrors = validationResult(req);
+    if (!inputErrors.isEmpty()) {
+      return res.status(422).json({ inputErrors: inputErrors.array() });
+    }
+
     await UserService.createUser(req.body);
     return res.status(200).send('User has been created successfully!');
   } catch (err) {
-    return res.status(500).send('Something wrong when tried create a user!');
+    return res.status(500).send('Something wrong when tried to create an user!');
   }
 };
 
 const deleteUser = async (req, res) => {
   try {
+    const inputErrors = validationResult(req);
+    if (!inputErrors.isEmpty()) {
+      return res.status(422).json({ inputErrors: inputErrors.array() });
+    }
+
     const userDeleted = await UserService.deleteUser(req.params.userId);
     if (userDeleted !== null) {
       return res.status(200).send('User has been deleted!');
@@ -62,8 +84,56 @@ const deleteUser = async (req, res) => {
 
     return res.status(404).send('User can not be deleted because the user does not exist!');
   } catch (err) {
-    return res.status(500).send('Something wrong when tried delete a user!');
+    return res.status(500).send('Something wrong when tried to delete an user!');
   }
 };
 
-export { validation, getUserById, getUserByCpf, createUser, deleteUser };
+const updateUser = async (req, res) => {
+  try {
+    const inputErrors = validationResult(req);
+    if (!inputErrors.isEmpty()) {
+      return res.status(422).json({ inputErrors: inputErrors.array() });
+    }
+
+    const userUpdated = await UserService.updateUser(req.params.userId);
+    if (userUpdated !== null) {
+      return res.send(200).send('User has been updated.');
+    }
+
+    return res.status(404).send('User can not be updated because the user does not exist!');
+  } catch (e) {
+    return res.status(500).send('Something wrong when tried to update an user!');
+  }
+};
+
+const authenticate = async (req, res) => {
+  try {
+    const inputErrors = validationResult(req);
+    if (!inputErrors.isEmpty()) {
+      return res.status(422).json({ inputErrors: inputErrors.array() });
+    }
+
+    const { cpf, password } = req.body;
+    const user = await UserService.getUserByCpf(cpf);
+    if (!user) {
+      return res.status(404).send('User with this cpf not found!');
+    }
+    const matchPassword = await bcrypt.compare(password, user.password);
+    if (matchPassword) {
+      const payload = {
+        cpf: user.cpf,
+        id: user.id
+      };
+      const token = await generateToken(payload);
+      console.log('User successfully authenticate!');
+      return res.status(200).send(token);
+    } else {
+      console.log('Wrong Password!');
+      return res.status(400).send('Wrong Password!');
+    }
+  } catch (e) {
+    return res.status(500).send('Something wrong when tried to authenticate an user!');
+  }
+};
+
+export { getUserById, getUserByCpf, createUser, deleteUser, updateUser, authenticate, getAllUsers };
